pyproject.toml
[tool.poetry]
name = "warehouse"
version = "0.0.0"
description = "Warehouse is the backend api for venera."
authors = ["VincentRPS"]

[tool.isort]
profile = 'black'

[tool.black]
target-version = ['py310']
skip-string-normalization = true

[tool.pyright]
reportMissingTypeStubs = false
reportPrivateUsage = false

[tool.poetry.dependencies]
python = "^3.10"
cassandra-driver = "^3.25.0"
bcrypt = "^3.2.2"
fastapi = "^0.78.0"
pydantic = "^1.9.1"
itsdangerous = "^2.1.2"
orjson = "^3.7.3"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
isort = "^5.10.1"
pyright = "^1.1.255"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
